requirements.txt
discord.py==2.5.2
openai==1.68.2
pytest==8.3.5